+++ apps/client/src/app/components/login-with-access-token-dialog/login-with-access-token-dialog.module.ts
@@ -9,14 +9,14 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { IonIcon } from '@ionic/angular/standalone';
 
-import { GfDialogHeaderModule } from '../dialog-header/dialog-header.module';
+import { GfDialogHeaderComponent } from '../dialog-header/dialog-header.component';
 import { LoginWithAccessTokenDialog } from './login-with-access-token-dialog.component';
 
 @NgModule({
   declarations: [LoginWithAccessTokenDialog],
   imports: [
     CommonModule,
-    GfDialogHeaderModule,
+    GfDialogHeaderComponent,
     IonIcon,
     MatButtonModule,
     MatCheckboxModule,